+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # `cdshealpix` Change Log
 
+
+
+--------------------------------------------------------------------------------
+
+## 0.1.3
+
+Released 2019-xx-xx.
+
+### Bug correction
+
+* fix BMOC logical operations (`not`, `and`, `or`, `xor`)
+
+### Added
+
+* `BMOCBuilderFixedDepth` to build a BMOC from a list of cells at the MOC depth
+* add sub-direcotry containing tests on MOCs (in a sub-module not to add the
+  `serde` and `serde-json` dependencies in the man crate) 
+
+### WARNING
+
+* BMOC logical operators still to be tested with `is_full` flags possibly set to `false`
+
 --------------------------------------------------------------------------------
 
 ## 0.1.2
@@ -13,7 +35,7 @@ Released 2019-01-24.
 
 ### Bug correction
 
-* `is_flag` always set to false in BMOC 
+* `is_flag` always set to false in BMOC now fixed
 
 ### Added